+++ backend/application/port/encryptor.go
@@ -1,0 +1,7 @@
+package port
+
+// TokenEncryptor defines operations for encrypting and decrypting tokens.
+type TokenEncryptor interface {
+	Encrypt(plaintext string) (string, error)
+	Decrypt(ciphertext string) (string, error)
+}
