+++ backend/application/port/github_repo.go
@@ -1,4 +1,4 @@
-package application
+package port
 
 import (
 	"context"
@@ -14,15 +14,4 @@ type GitHubRepoRepository interface {
 	FindByChannelID(ctx context.Context, channelID string) (*entities.GitHubRepo, error)
 	Delete(ctx context.Context, repoID string) error
 	UpdateToken(ctx context.Context, repoID string, encryptedToken string) error
-}
-
-// TokenEncryptor defines operations for encrypting and decrypting tokens.
-type TokenEncryptor interface {
-	Encrypt(plaintext string) (string, error)
-	Decrypt(ciphertext string) (string, error)
-}
-
-// GitHubIssueCreator defines the operation for creating GitHub Issues via the GitHub API.
-type GitHubIssueCreator interface {
-	CreateIssue(ctx context.Context, owner, repo, token, title, body string) (issueURL string, err error)
 }
